backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
python-dotenv==1.0.0
openai==1.3.7
httpx==0.27.2
pypdf==3.17.4
python-docx==1.1.0
reportlab==4.0.7
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.1
Pillow==12.1.1
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
drf-spectacular==0.27.0
